Started formatting class data
Edited the example JSON to be more usable for me, and started the
process of formatting that data into a datastructure. I am using a
DIGraph for this so I can use topological sorting. It is more than
likely not going to work at the current moment, there's a few kinks I
still have to figure out.